requirements.txt
python-telegram-bot==22.1
openai
python-dotenv
pytesseract
Pillow
Flask